moved previous implementation of factory method example into a separate .dart file & revised script.dart
